refactor: more cleanup and renaming
