--- src/classes/ThemeActions.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Vibrato\Classes;
-
-class ThemeActions
-{
-    public function init()
-    {
-    }
-}
